make state changes
