+++ src/app/static/src/tests/mocks.ts
@@ -27,7 +27,7 @@ import {initialModelRunState, ModelRunState} from "../app/store/modelRun/modelRu
 import {emptyState, RootState} from "../app/root";
 import {initialStepperState, StepperState} from "../app/store/stepper/stepper";
 import {initialMetadataState, MetadataState} from "../app/store/metadata/metadata";
-import {initialLoadState, LoadState} from "../app/store/load/load";
+import {initialLoadState, LoadState} from "../app/store/load/state";
 import {initialModelOptionsState, ModelOptionsState} from "../app/store/modelOptions/modelOptions";
 import {initialModelOutputState, ModelOutputState} from "../app/store/modelOutput/modelOutput";
 import {
